+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Enhancements
 
 ### Bug Fixes
+* Render RRF variants correctly in VariantDetailsModal ([#836](https://github.com/opensearch-project/dashboards-search-relevance/pull/836))
 
 ### Infrastructure